Move generator scripts to tools/ folder
