--- css.js
@@ -1,2 +1,0 @@
-import './';
-import './styles/styles.scss';
