+++ tools/css.js
@@ -1,0 +1,4 @@
+// This script is used to help webpack watch the folders
+
+import '../';
+import '../styles/styles.scss';
