+++ tools/initFolders.js
